Fix another post-shutdown shutdown-blocking task
This addresses a fairly subtle Storage Service issue as detailed in the
new code comments. While the issue turns out to be harmless for users
(it doesn't affect which data we do or do not persist on shutdown), it
does result in noisy but ultimately suppressed DCHECKs when running many
integration test suites, including web tests.

Fixed: 1174179
Change-Id: I4defe8c3b57572381c35e6edcbb0317a4b941e44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2703271
Commit-Queue: Ken Rockot <rockot@google.com>
Reviewed-by: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855401}