Prevent same-document navigations without document loaded after crash
The same-document history navigation code will not try to perform a
same-document navigation if IsRenderFrameLive() is false. However that
becomes true after a crash as soon as the process for the SiteInstance
comes back, even though the frame itself was still crashed. Instead, we
should check for the frame itself having a document in it which we can
do by checking has_committed_any_navigation(). This becomes false on
process exit and true after the first commit _for that frame_. This
means the SiteInstance coming back alive can't cause a RenderFrameHost
to try mutate a renderer frame that doesn't exist.

This adds 4 tests to verify already working behaviour as well as the
change:
1. Same-document regular navigations after crash and after session
   restore
2. Same-document history navigations after crash and after session
   restore

R=avi@chromium.org

Bug: 1169423
Change-Id: I8f5b3b8108c63c9aa9977030dead9ef6e8a68d34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2702837
Commit-Queue: danakj <danakj@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855415}