Handle DOM-created tables with atypical structure
Fall back on normal AddNodeChildren() for atypical tables where
AddTableChildren() doesn't find all the children.

Normally, when an HTML <table> is parsed, it automatically gets
at least a <tbody> section, even if this was skipped in the markup.
The AddTableChildren() method exists to  ensure that table sections
are added in the order they are rendered visually, enforcing the order
of caption > thead > tbodies > tfoot, without having to traverse layout.
Unfortunately, AddTableChildren() assumes a typical table structure,
and misses children that are not one of the above.

The new code switches off the specialized table handling unless it
has the typical structure.

A good future change would be to update the table roles from
layoutTable, layoutTableRow, etc. when there is enough new content to
show it should be treated as a data table.

Bug: 1179516
Change-Id: I1b40eac5caeaaae68bef34c39192e561c1757716
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2702659
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: David Tseng <dtseng@chromium.org>
Reviewed-by: Nektarios Paisios <nektar@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855460}