[Switch Access] Natural order for scan speed slider UX
Currently, the Switch Access scan speed sliders work such that
the farther to the right, the slower the speed. As we're introducing
soon a third slider that will work the other way, it would be good to
align the first two sliders so that the farther to the right, the
faster the speed.

After: https://screenshot.googleplex.com/4NpedGDU9QdSNNZ

Bug: 1179531
Change-Id: Ie7029c21e312ad77644a4e95c6e7a1b09d000b1b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2702900
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: Anastasia Helfinstein <anastasi@google.com>
Commit-Queue: Josiah Krutz <josiahk@google.com>
Cr-Commit-Position: refs/heads/master@{#855467}