WebUI: Delete unused test file
This test file is using old, deprecated pre JS modules code, and does
not appear to be included or run anywhere. Remaining contents of this
folder seem to already be migrated to JS modules, so if testing for
the code in this folder is needed, new JS module based tests should be
added instead.

Bug: 1173575
Change-Id: I1020ba0eceeee445ee9c10d3f06583248a27da85
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2699024
Commit-Queue: Rebekah Potter <rbpotter@chromium.org>
Reviewed-by: dpapad <dpapad@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855513}