Use 'SEARCH' instead of 'META' for the ChromeOS 'WIN' VKEY
This CL updates the ChromeOS keyboard layout monitor to map 'SEARCH'
to the right and left 'WIN' keys for the layout provided to the web
client.  This means that mobile-web users connecting to ChromeOS
hosts will see the magnifying glass icon instead of the Windows
logo for this key.

Though this key is also called the 'launcher' or 'everything' key,
a quick internet search shows the search icon more often than the
icon for the other usages.  Searching though Chromium for these
terms brings up 'search' as the term most often used to refer to
this key.

Bug: 1175857
Change-Id: I2b628735a075d77c6b0a83014c15df78ae287db5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2704729
Commit-Queue: Jamie Walch <jamiewalch@chromium.org>
Auto-Submit: Joe Downing <joedow@chromium.org>
Reviewed-by: Jamie Walch <jamiewalch@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855543}