[Extensions] Re-enable a pageCapture API test.
The SaveAsMHTMLWithoutFileAccess was disabled for flakiness but I was
unable to find consistent evidence of this flakiness looking at the
historic runs on the flakes portal, using the link supplied in the bug.
I was also unable to find consistent evidence of timeouts per the bug
description.

I ran the test for 3000 iterations on the CQ bots, including all of
the MSAN and ASAN configurations. Most configurations completed all
3000 iterations without a single failure. Some Linux-based bots had
a single crash caused by a known shared library loading issue that's
unrelated to the test.

Patchset 1 contains the results of running the tests as a legacy
background page-based extension. Patchset 2 contains the results
of the running the tests as a Service Worker-based extensions.

Bug: 1156323
Change-Id: I66a5db99b290ed213c6855c019fab353b0ab429d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2699014
Auto-Submit: David Bertoni <dbertoni@chromium.org>
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855562}