[Switch Access] Top-align back button with element
Because of issues with the focus rings overlapping the back button, as
well as an overall redesign of the UX, the back button should now be
top-aligned with the element being focused, instead of appearing off the
corner.

AX-Relnotes: The Switch Access back button is now next to the group.
Bug: 1152610
Change-Id: I323139c76d49077ae9c67035cb62a8174fbbb854
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2691537
Commit-Queue: Anastasia Helfinstein <anastasi@google.com>
Reviewed-by: David Tseng <dtseng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855567}