ozone: clipboard: Support nested requests
So far, ClipboardOzone implementation has been assuming nested clipboard
request shouldn't happen, which is actually a wrong assumption (see
crbug.com/1165466 for further details).

This change refactors ClipboardOzone's Request implementation to be a
template-based class which encapsulates the nested loop manipulation and
no longer assumes that a single request is processed at a given time. To
fully fix the issue, Ozone backends must also be able to handle nested
requests, which is currently not true for Wayland, for example, so this
CL also adds a TODO to the WaylandClipboard class, which will be
addressed as follow-ups.

Also, as discussed in [1], more follow-ups are expected, which include
changes in the PlatformClipboard API as well as in both Wayland and X11
backends, in order to move the sync logic (ie: nested loops, etc) down
to the Ozone backend.

[1] https://bugs.chromium.org/p/chromium/issues/detail?id=1165466#c11

R=dcheng@chromium.org

Bug: 1165466
Test: Tested manually with subsequent pastes into webcontents text input items.
Change-Id: I2eb65184e46ba160117d7c9bbb2a0d87b4b40269
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2676322
Commit-Queue: Nick Yamane <nickdiego@igalia.com>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855582}