Remove obsolete DCHECK in ComputeOffsetMapping
This DCHECK was very useful when the code was not very stable.
These days it only catches rare missing or additional ZWSP.
The ZWSP cases is still nice to fix, but as we watch reports,
we are not seeing any crashes or issues caused by the
difference at this point.

We kept the DCHECK to catch other possible breakages, but it
looks like it is causing some troubles to our developers.
The risk of future possible breakage doesn't look high enough
to justify the trouble, hence removing.

Bug: 1022422
Change-Id: I35d76be259fa3008ef63a3cd062caf41ad535462
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705218
Commit-Queue: Kent Tamura <tkent@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Auto-Submit: Koji Ishii <kojii@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855586}