[MBI] Apply current policy logic to per-AGS CompositorTaskRunner
Before this CL, per-AGS (Agent Scheduling Group) CompositorTaskRunner
was just an alias of MainThreadSchedulerImpl::CompositorTaskRunner().

After this CL, if the kMbiCompositorTaskRunnerPerAgentSchedulingGroup
is enabled, per-AGS CompositorTaskRunner is actually used. Also, the
compositor related policy logic is applied to per-AGS
CompositorTaskRunners.

[IdleTimeEstimator]

Previously, IdleTimeEstimator was using a single compositor task queue
to measure the expected idle duration. But this time, the number of
compositor task queues will change over time. So this CL adapted to
it.

[Policy]

The policy around the compositor task queue stays the same as before
in this CL.

Bug: 1105403
Change-Id: Ib587498dcc654b3c80ff024e36cae6ce6253cf40
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2658697
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Reviewed-by: Katie Dillon <kdillon@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855597}