Crostini: give a reason why Crostini is not available
In addition to providing Yes/No answer to IsCrostiniEnabled DBus query
also provide a reason why it is disallowed, so that we can show this
information to the user in crosh/vmc.

Because the intended user of the strings is crosh/vmc which is not going
to be localized, we are not attempting l10n here.

Bug: b/179688287
Change-Id: I9387328823c19a9d0830672ede1b2d96bb9afeb7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2689037
Commit-Queue: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>
Reviewed-by: Joel Hockey <joelhockey@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855626}