WebXR depth: handle a case where depth is both optional and required
When depth sensing was requested both as optional and required feature,
the ArCoreImpl would successfully create a session even in cases where
depth is not supported by ARCore - this is now fixed.

Additionally, add verification on the browser side to ensure that a
runtime does not report that session creation succeeded but not all
required features have been enabled.

Fixed: 1171464,1171459
Change-Id: I62143f50ffd7ee8f55a7ae746fb4668d1c7ad549
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705409
Commit-Queue: Piotr Bialecki <bialpio@chromium.org>
Reviewed-by: Alexander Cooper <alcooper@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855632}