[Autofill] Fixed clearing of autocomplete, particularly in shadow DOM.
Autocompleted (not: autofilled) fields in a shadow DOM were not cleared.
Since crrev.com/c/2676625, this extends to fields in the light DOM.

The original cause, prior to crrev.com/c/2676625, was that
ClearPreviewedFormWithElement() attempted to clear an element, if that
element is formless, by retrieving all formless elements and clearing
them. This retrieval does not enter shadow DOMs. A formless field in a
shadow DOM was therefore never cleared.

With crrev.com/c/2676625, the bug was even extended to the light DOM
because DoPreviewFieldWithValue(), which is called by autocomplete, did
not add the previewed field to |previewed_elements_|.

Bug: 973018
Change-Id: Ia0189c102c6ebd4913e8407c26d642d9b4e15f61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2703969
Reviewed-by: Matthias Körber <koerber@google.com>
Commit-Queue: Matthias Körber <koerber@google.com>
Auto-Submit: Christoph Schwering <schwering@google.com>
Cr-Commit-Position: refs/heads/master@{#855682}