Disable layout test
This test uses the old runtime infra to lookup an extension we are
trying to migrate in https://crrev.com/c/2703575. We need to disable
the test to allow for the CL with the migration to land before we update
and re enable it.

Bug: 1134103
Change-Id: I0038cf722cd12db03ae5feb69bc5a2f7f13d19f3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2704903
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855693}