[ozone/x11] Enabled X11-specific timeout in Vulkan swap chain.
On Linux/X11, the Vulkan swap chain may hang in certain normal use cases
such as screen turned off.  To work around that, we set a timeout when
calling vkAcquireNextImageKHR so it returns an error in those use cases.

This CL enables that logic for Ozone/X11 by extracting the timeout to
the level of VulkanSurface, where it can be set by platform
implementations.

Bug: 1098237
Change-Id: Icaa7ca7884680bf4255b9b0ab47dc1958f79f014
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705382
Reviewed-by: Peng Huang <penghuang@chromium.org>
Commit-Queue: Alexander Dunaev <adunaev@igalia.com>
Cr-Commit-Position: refs/heads/master@{#855713}