Add Bundle URL to requests to resources in WebBundle
Before this patch, a request to a resource in WebBundle only had the
resource's URL and unguessable bundle token. It was problematic in
subframe navigation to a urn:uuid URL, because request has no
information about the bundle's origin. It caused
NavigationRequest::GetWebBundleURL() return an empty URL, and resulted
in a CHECK failure in NavigationRequest::GetOriginForURLLoaderFactory().

To fix this issue, this patch adds bundle_url field to
WebBundleTokenParams (which is part of URLRequest). This field is
populated by renderer process, and verified by the Network Service.
NavigationRequest::GetWebBundleURL() refers this field instead of
web_bundle_url field of URLResponseHead.

Bug: 1179093
Change-Id: I6bfc7374e2fcde44c4153fc21bfafb5510a0c81d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2692440
Reviewed-by: Hayato Ito <hayato@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Tsuyoshi Horo <horo@chromium.org>
Commit-Queue: Kunihiko Sakamoto <ksakamoto@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855727}