[profiles] Remove lazy entry init in ProfileInfoCache
ProfileAttributesEntries are currently laziliy initialized in
GetProfileAttributesWithPath(). However, all the entries get
initialized right away in ProfileInfoCache's constructor on all
platforms but Android due to LoadGAIAPictureIfNeeded() call.

This leads to a weird implementation of GetNumberOfProfiles() which
assumes that entries are initialized on non-Android platforms and has a
special path for Android.

This CL removes the lazy initialization of profiles attributes entries
in ProfileInfoCache. Now, the entries are initialized immediately after
a profile being added which makes the code simpler.

Bug: 1175835
Change-Id: Ie959e687ff4e5de44949ebbc5f38270e645e261c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2695827
Reviewed-by: Kush Sinha <sinhak@chromium.org>
Reviewed-by: Mihai Sardarescu <msarda@chromium.org>
Commit-Queue: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855743}