Remove NOTREACHED in CanvasRenderingContext::Reshape
The NOTREACHED check is superfluous because the method does not need to
be implemented in derived classes.  In the WebGPU case, for example,
GPUSwapChainProvider pulls the texture size from context host each time
and no internal state changes are necessary when the canvas is resized.
Therefore, it makes senes to let the Reshape method be unimplemented.

BUG=1173676

Change-Id: I0b851a8af6963fbaad9cc2cba38216c780c9b0fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705152
Reviewed-by: Fernando Serboncini <fserb@chromium.org>
Commit-Queue: Justin Novosad <junov@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855746}