Turn on cancellations for all builders by extending grace period to 2min
Previously, cancellation may drop caches on the bot(crbug/1127113), thus
we have to disable cancellations on several builders. go/crrv/c/2689801
has proved on mac-rel that 2min grace_period can save almost 100% of
caches on cancelled builds.

It makes sense to expand it to all builders, because if the builders
are light and never hit the grace period, this CL has no impact on
them.

As a result, we could safely re-enable cancellations for all builders.

BUG=chromium:1179836
TEST=verification on mac-rel

Change-Id: Ia7718e49287f31aee735223ce6ebb39420fc83a4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2704957
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Commit-Queue: Xinan Lin <linxinan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855750}