Fix ShapeResult::IsStartSafeToBreak when no glyphs
This patch fixes |ShapeResult::IsStartSafeToBreak| when the
first character does not have a glyph.

There are two cases it can occur:
1. Shaped across elements, then |SubRange| split the
   |ShapeResult| at the middle of a cluster or a ligature.
2. The first character did not produce any glyphs.

In case 1, |ShapeResult::CopyRange| produces a |RunInfo|
without any |HarfBuzzRunGlyphData|. Such |RunInfo| is not
useful nor necessary that this patch stops producing them.

With the fix, if the glyph for the first character is missing,
|IsStartSafeToBreak| should be |false|, because the middle of
a cluster or a ligature is not safe-to-break.

Case 2 is unlikely because we try hard to get the last resort
font, but it could still occur. It may be safe, or may be not,
we cannot determine without glyph data. This patch changes
both cases to return |false|.

Bug: 1179678
Change-Id: I55b61e180ba1306a376a40dd3f31a22fe1913676
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705658
Commit-Queue: Koji Ishii <kojii@chromium.org>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855779}