Make dedicated worker's COEP reporter
This CL adds worker's COEP reporter and update the raw pointer of
ancestor's COEP reporter to a weak pointer.

Without PlzDedicatedWorker, no behavior changes. Reports are sent to
the ancestor's COEP reporter (same as the current behavior) but it's
not align with the spec. When the ancestor's COEP reporter is used, the
existence of it is checked to make sure the frame is not destructed.

With PlzDedicatedWorker, reports are sent to the creator's COEP reporter
if a worker's script loading fails and sent to the worker's COEP
reporter if fetch inside a worker fails.

Bug: 1175436, 1060837
Change-Id: I2ea6b2d51d381a283b5cc722e4517cfc842967c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2692088
Commit-Queue: Asami Doi <asamidoi@chromium.org>
Reviewed-by: Matt Falkenhagen <falken@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Reviewed-by: Yutaka Hirano <yhirano@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855816}