[feedback] Remove unused getters / #includes
Those getters / #includes are dependencies on //content that
would not work on iOS. Remove them as they are not needed (make
it easier to fix the code to build on iOS).

This does not remove all dependencies on //content. This will
happen in followup CLs, but //components/feedback classes no
longer exposes public getter to access //content objects.

Bug: 1179672
Change-Id: I4ab32ebca73d7ed5fbdfd25f1c3bb316eca2744b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2700394
Commit-Queue: Sylvain Defresne <sdefresne@chromium.org>
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855825}