vaapi: Fix H264 CENCv1 protected content
There was an issue with CENCv1 H264 content when there were encrypted
SEI NALUs in the stream. Previously, we were just ignoring those NALUs.
Now we save those NALUs in the decoder and their subsamples, so that
when we need to process an encrypted slice header preceded by SEI NALUs
in the same sample, we can pass them all into the accelerator so it can
properly decrypt the entire sample block.

BUG=b:179811973
TEST=Netflix and Widevine H264 CENCv1 content plays back

Change-Id: I27d2894a3e7f5f70ae49dc5e2ad93ba595545587
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2705279
Reviewed-by: Hirokazu Honda <hiroh@chromium.org>
Reviewed-by: Jao-ke Chin-Lee <jchinlee@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Cr-Commit-Position: refs/heads/master@{#855827}