Notify user activation for local frame when resuming playback if needed
On crrev.com/c/2593109 we migrated MediaPlayerDelegateMsg_RequestPlay
to Mojo but we made a mistake and forgot to include a call to
LocalFrame::NotifyUserActivation() when media playback was being
resumed in the HTMLMediaElement as the result of such message being
received from the browser process, leading to bug 1171124.

To fix this, let's re-arrange the code on HTMLMediaElement so that
user activation gets notified in the exact same scenario as before.

Bug: 1171124, 1039252
Change-Id: If4bd7bebc8ad3521cd2870e99a631cc65877eb8f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2662980
Commit-Queue: Mario Sanchez Prada <mario@igalia.com>
Commit-Queue: Mounir Lamouri <mlamouri@chromium.org>
Auto-Submit: Mario Sanchez Prada <mario@igalia.com>
Reviewed-by: Mounir Lamouri <mlamouri@chromium.org>
Reviewed-by: Chrome Cunningham <chcunningham@chromium.org>
Cr-Commit-Position: refs/heads/master@{#855837}